requirements.txt
streamlit==1.28.1
python-docx==0.8.11
pandas==2.0.3
numpy==1.24.3
PyYAML==6.0
requests==2.31.0
Pillow==10.0.0
